+++ R/shared-secret-filter.R
@@ -5,7 +5,7 @@ sharedSecretFilter <- function(req, res){
     supplied <- req$HTTP_PLUMBER_SHARED_SECRET
     if (!identical(supplied, secret)){
       res$status <- 400
-      stop("The provided shared secret did not match expected secret.")
+      return(list(error = "Shared secret mismatch."))
     }
   }